Add a default export
When importing n3 from an ES module in Node, the exports will be
treated as a default export, because the package itself is not an
ES module.

This _actually_ exports that default module, so that the same
syntax (i.e. importing the default export) can be used with both
Node and bundlers.